QazSign-Mobile/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "gunicorn>=24.1.1",
    "mediapipe>=0.10.32",
    "numpy>=2.4.1",
    "opencv-python-headless>=4.13.0.90",
    "pillow>=12.1.0",
    "tensorflow>=2.20.0",
]
